Moved functions to lib folder to cleanup project structure.
