+++ lib/isEmpty.js
