+++ lib/isEqual.js
