+++ lib/isEven.js
