+++ lib/isOdd.js
